+++ tiledb/sm/cpp_api/tiledb_experimental
@@ -45,5 +45,6 @@
 #include "query_condition_experimental.h"
 #include "query_experimental.h"
 #include "subarray_experimental.h"
+#include "vfs_experimental.h"
 
 #endif  // TILEDB_EXPERIMENTAL_CPP_H
